created appContext and add in comboboxvalue
