+++ frontend/src/components/Conference.jsx
@@ -1,10 +1,4 @@
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-import useGetAccessToken from "../custom_hooks/useGetAccessToken";
-import { getConferences } from "../services/conferences";
-import { useState } from "react";
-import { NormalComboBox } from "./NormalComboBox";
 import { countries } from "../utils/countries";
-import { useLocation } from "react-router-dom";
 import {
   Form,
   FormControl,
@@ -26,6 +20,7 @@ import DatePicker from "./DatePicker";
 import { useForm, useFieldArray } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as z from "zod";
+import { useAppContext } from "../context/appContext";
 const Conference = () => {
   const FormSchema = z
     .object({
@@ -83,30 +78,9 @@ const Conference = () => {
     });
   };
   const { toast } = useToast();
-  const location = useLocation();
-  const [comboBoxValue, setComboBoxValue] = useState(location.state);
-  const getAccessToken = useGetAccessToken();
-
-  const { data: conferences, isLoading: isConferenceFetching } = useQuery({
-    queryKey: ["conferences"],
-    queryFn: async () => {
-      const accessToken = await getAccessToken();
-      return getConferences(accessToken);
-    },
-    refetchOnWindowFocus: false, // it is not necessary to keep refetching
-  });
+  const { comboBoxValue } = useAppContext();
   return (
     <div className="flex flex-col w-full p-12">
-      {!isConferenceFetching && (
-        <NormalComboBox
-          options={conferences}
-          validateProperty={"name"}
-          displayProperty={"name"}
-          fieldName={"conference"}
-          value={comboBoxValue}
-          setValue={setComboBoxValue}
-        />
-      )}
       <h1 className="text-4xl font-bold">{comboBoxValue}</h1>
       <div className="flex w-full gap-6">
         <div className="w-full flex p-6 border-[#EAECF0] flex-col gap-6  shadow-md">
