+++ frontend/src/components/DashboardLayout.jsx
@@ -1,4 +1,10 @@
 import React, { useState, useEffect } from "react";
+import { useQuery, useQueryClient } from "@tanstack/react-query";
+import useGetAccessToken from "../custom_hooks/useGetAccessToken";
+import { getConferences } from "../services/conferences";
+
+import { NormalComboBox } from "./NormalComboBox";
+
 import { ProfileIcon } from "./ProfileIcon";
 import {
   DropdownMenu,
@@ -10,19 +16,43 @@ import { Home, BarChart2, CheckSquare, Flag } from "lucide-react";
 import { Outlet, useLocation } from "react-router-dom";
 import { cn } from "../lib/utils";
 import { useAuth0 } from "@auth0/auth0-react";
+import { useAppContext } from "../context/appContext";
 const DashboardLayout = () => {
   const { pathname } = useLocation();
   const { logout, user } = useAuth0();
 
   const [userName, setUserName] = useState("");
+  const location = useLocation();
+  const { comboBoxValue, setComboBoxValue } = useAppContext();
+  const getAccessToken = useGetAccessToken();
 
+  const { data: conferences, isLoading: isConferenceFetching } = useQuery({
+    queryKey: ["conferences"],
+    queryFn: async () => {
+      const accessToken = await getAccessToken();
+      return getConferences(accessToken);
+    },
+    refetchOnWindowFocus: false, // it is not necessary to keep refetching
+  });
   useEffect(() => {
     setUserName(user.name);
   }, [user]);
 
   return (
     <div className="flex flex-col min-h-screen layout">
-      <div className="pt-4 pr-6 border bottom-2">
+      <div className="pt-4 pr-6 border ablublu bottom-2">
+        <div>
+          {!isConferenceFetching && (
+            <NormalComboBox
+              options={conferences}
+              validateProperty={"name"}
+              displayProperty={"name"}
+              fieldName={"conference"}
+              value={comboBoxValue}
+              setValue={setComboBoxValue}
+            />
+          )}
+        </div>
         <div className="flex items-center justify-end gap-2 mb-4 ">
           <p className="text-color">{userName}</p>
           <DropdownMenu>
